Moved Classes to correct structure school/faang/BJS2_68344
